Move generate.tsx up
